refactor: API clientの責務を分離
